chore: rename `statetest` to `validatorstest`
