+++ snow/validators/validatorstest/state.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2019-2024, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package statetest
+package validatorstest
 
 import (
 	"context"
